Make card size in Decklist relative to DecklistPanel's size
Cards are no longer tiny.
Just making the size dependent on DecklistPanel's size unfortunately led
to zero-sized cards on startup as the Panel's bounds are not yet
initialized when the first decklist is leaded.
To work around this a LayoutUpdated handler is added to DecklistPanel
to load the Decklist and remove itself since at that point bounds are
set up