Make whole width hoverable in ListBoxes
Previously hovering over an entry in a ListBox would only work if the mouse was over the text because the PointerEnter event was registered on the TextBlock which would not fill the ListBoxItem's space.
Now the event is registered on a Border with transparent background containing the TextBlock (huge thanks to timunie from the Avalonia Matrix/Telegram channel, I would not have found that out).
The Border will only hit-test if it has any background, even if it is transparent (apparently null indicates a non-open window)